requirements.txt
websocket-client==1.4.2
requests==2.28.1
